观察者模式-基于Sping状态机-end

+++ src/main/java/com/coderwhs/designPattern/model/entity/Order.java
@@ -3,6 +3,8 @@ package com.coderwhs.designPattern.model.entity;
 import com.coderwhs.designPattern.model.enums.OrderStateEnum;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -12,6 +14,6 @@ public class Order {
     private String orderId;
     private String productId;
     private OrderStateEnum orderState;//订单状态
-    private Float price;//商品价格
+    private BigDecimal price;//商品价格
     private String userId;//当前用户唯一Id
 }
